--- main.py
@@ -1,10 +1,0 @@
-from website import create_app
-
-app = create_app()
-
-if __name__ == '__main__':
-    app.run(debug=True)
-
-
-if __name__ == "__main__":
-    app.run(debug=True, port=5000)
