+++ website/templates/home.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Login</title>
     <link rel="stylesheet" href="../static/style.css">
-    <script src="../static/myscripts.js"></script>
+    <script src="myscripts.js"></script>
 </head>
 <body>
     <nav>
@@ -14,10 +14,10 @@
             <li><a href="#">About</a></li>
             <li><a href="#">Make your appointment</a></li>
             <li><a href="/login">Login</a></li>
-            <li><a href="/signup">Sign Up</a></li>
+            <li><a href="signup.html">Sign Up</a></li>
         </ul>
     </nav>
-    
+    n
     <div class="footer-wrapper">
         <footer class="footer">
                 <div class="footer-col">
