+++ website/templates/login.html
@@ -5,16 +5,16 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Login</title>
     <link rel="stylesheet" href="../static/style.css">
-    <script src="../static/myscripts.js"></script>
+    <script src="myscripts.js"></script>
 </head>
 <body>
     <nav>
         <ul>
-            <li><a href="/" class="active">Home</a></li>
+            <li><a href="home.html" class="active">Home</a></li>
             <li><a href="#">About</a></li>
             <li><a href="#">Make your appointment</a></li>
-            <li><a href="/login">Login</a></li>
-            <li><a href="/signup">Sign Up</a></li>
+            <li><a href="login.html">Login</a></li>
+            <li><a href="signup.html">Sign Up</a></li>
         </ul>
     </nav>
